+++ classes/jigoshop_options.class.php
@@ -470,6 +470,17 @@ class Jigoshop_Options implements Jigoshop_Options_Interface {
 				'type' => 'single_select_country',
 			),
 			array(
+				'name' => __('Default Country/Region for customer', 'jigoshop'),
+				'desc' => '',
+				'tip' => __('This is the country for your clients with new accounts.', 'jigoshop'),
+				'id' => 'jigoshop_default_country_for_customer',
+				'std' => $this->get_option('jigoshop_default_country'),
+				'type' => 'single_select_country',
+				'options' => array(
+					'add_empty' => true,
+				),
+			),
+			array(
 				'name' => __('Currency', 'jigoshop'),
 				'desc' => '',
 				'tip' => __('This controls what currency the prices are listed with in the Catalog, and which currency PayPal, and other gateways, will take payments in.', 'jigoshop'),